% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\title{Filters dataset using a year range and removes NA values}
\usage{
clean_data(data, col, from, to)
}
\arguments{
\item{data}{Specified dataframe}

\item{col}{Specified year column to filter on}

\item{from}{Starting year}

\item{to}{Ending year}
}
\value{
A filtered dataframe
}
\description{
Filters dataset using a year range and removes NA values
}
\examples{
#clean_data(df,Year,2010,2019)
}
