% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season_filter.R
\name{season_filter}
\alias{season_filter}
\title{Filters dataset using a year range}
\usage{
season_filter(data, col, from, to)
}
\arguments{
\item{data}{Specified dataframe}

\item{col}{Specified year column to filter on}

\item{from}{Starting year}

\item{to}{Ending year}
}
\value{
A filtered dataframe
}
\description{
Filters dataset using a year range
}
\examples{
select_data <- season_filter(data, Year, 2011, 2015)
}
