% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_to_from.R
\name{select_to_from}
\alias{select_to_from}
\title{Filters dataset using a specified range on specific column}
\usage{
select_to_from(data, col, from, to)
}
\arguments{
\item{data}{Specified dataframe}

\item{col}{Specified variable to filter on}

\item{from}{Starting point}

\item{to}{Ending point}
}
\value{
A filtered dataframe
}
\description{
Filters dataset using a specified range on specific column
}
\examples{
select_to_from(mtcars,mpg,21,22.8)
}
