% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\name{split_data}
\alias{split_data}
\title{Splits dataset into a training and testing dataset and writes it to a CSV}
\usage{
split_data(data, proportion, specified_strata)
}
\arguments{
\item{data}{Specified dataframe}

\item{proportion}{Specified split proportion}

\item{specified_strata}{Specified variable that you are prediciting for}
}
\value{
function writes final dataframe into two dataframe
}
\description{
Splits dataset into a training and testing dataset and writes it to a CSV
}
\examples{
#split_data(dataset,0.7,"Is_All_Star")
}
