% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion.R
\name{proportion}
\alias{proportion}
\title{Return a proportion of a specified column}
\usage{
proportion(data, column)
}
\arguments{
\item{data}{Specified dataframe}

\item{column}{Specified column to apply function on}
}
\value{
A summary dataframe containing proportion of specified column in percentage
}
\description{
Return a proportion of a specified column
}
\examples{
proportion(mtcars,vs)
}
