% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\name{split_data}
\alias{split_data}
\title{Splits dataset into a training and testing dataset and saves it into 2 global variable}
\usage{
split_data(data, proportion, specified_strata)
}
\arguments{
\item{data}{Specified dataframe}

\item{proportion}{Specified split proportion}

\item{specified_strata}{Specified variable that you are prediciting for}
}
\value{
function saves final dataframe into two global variables "data_testing", "data_training"
}
\description{
Splits dataset into a training and testing dataset and saves it into 2 global variable
}
\examples{
split_data(mtcars,0.7,"vs")
}
