% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_knn.R
\name{visualize_knn}
\alias{visualize_knn}
\title{Visualizes a knn model dataframe}
\usage{
visualize_knn(data, x, y, title)
}
\arguments{
\item{data}{Specified dataframe}

\item{x}{Specified x-axis}

\item{y}{Specified y-axis}

\item{title}{Specified title of knn plot}
}
\value{
Returns a plot using specified dataframe
}
\description{
Visualizes a knn model dataframe
}
\examples{
#visualization <- visualize_knn(data_fit,x = height,y = weight,title = "Prediction")
}
